src/app/doubles/doubles.ts
export interface StringInfoInterface {
  lowerCase: string;
  upperCase: string;
  characters: string[];
  length: number;
  extraInfo?: Object;
}

type LoggerServiceCallback = (arg: string) => void;

export function calculateComplexity(stringInfo: StringInfoInterface) {
  return Object.keys(stringInfo.extraInfo).length * stringInfo.length;
}

export function toUpperCaseWithCb(
  arg: string,
  callBack: LoggerServiceCallback
) {
  if (!arg) {
    callBack("Invalid arguments!");
    return;
  }
  callBack(`called function with ${arg}`);
  return arg.toUpperCase();
}

export class OtherStringUtils {
  public toUpperCase(arg: string) {
    return arg.toUpperCase();
  }

  public logString(arg: string) {
    console.log(arg);
  }
}
